\name{weightQuant-package}
\alias{weightQuant-package}
\alias{weightQuant}
\docType{package}
\title{
Weights for incomplete longitudinal data and quantile regression
}
\description{
Functions for the estimation of observation-specific weights for
incomplete longitudinal data. A bootstrap method is also provided to
obtain standard erros of weighted quantile regressions.
}
\details{
\tabular{ll}{
Package: \tab weightQuant\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2018-03-14\cr
License: \tab GPL (>=2.0)\cr
}
}
\author{
Viviane Philipps
}
%\references{
%}
\keyword{ package }
